Microservices/monitoring/requirements.txt
fastapi
uvicorn
pymongo
pika
python-socketio
python-dotenv
scikit-learn
numpy
requests
pandas
